Make hostname canonical in /etc/hosts file
Instead of making the hostname an alias in the /etc/hosts file, make it
a canonical name (first name after the loopback address). RabbitMQ fails
to authenticate on hosts that have the hostname set as an alias (any
record after the canonical name in a line). Error thrown is:

  Hostname mismatch: node believes its host is different

Related to rabbitmq/discussions#53

Signed-off-by: Jason Rogena <jason@rogena.me>